core: Move lattice_switch to lb interface.
