+++ src/core/grid_based_algorithms/lattice.hpp
@@ -26,21 +26,10 @@
  * surrounds the local lattice sites.
  */
 
-#ifndef _LATTICE_HPP
-#define _LATTICE_HPP
+#ifndef CORE_LB_LATTICE_HPP
+#define CORE_LB_LATTICE_HPP
 
 #include "utils/Vector.hpp"
-
-/** Switch determining the type of lattice dynamics. A value of zero
- *  means that there is no lattice dynamics. Different types can be
- *  combined by or'ing the respective flags.
- *  So far, only \ref LATTICE_OFF and \ref LATTICE_LB exist.
- */
-extern int lattice_switch;
-
-#define LATTICE_LB 1     /** Lattice Boltzmann */
-#define LATTICE_LB_GPU 2 /** Lattice Boltzmann */
-#define LATTICE_OFF 0    /** Lattice off */
 
 class Lattice {
 public:
